Move ServerCtl into Fuseki1; use in Fuseki1 tests
